Changed the version check for the checkFormat/reformat tasks to JDK 16+.
It seems that the same version of the library run on JDK 11 doesn't support records (and other post 11 features) whereas it does when run on JDK 16.  Because we now have records in the code (even though it's just in the tests), we can no longer run the format tasks on JDK 11 as they will fail.